予想/ai-auto-mailer/boatrace-dashboard/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
psycopg2-binary>=2.9.9
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyjpboatrace>=0.4.0
